add trie removal function
Adds the remove() function to the Trie data structure class. Associates end_symbol with a numeric value
used to represent the number of instances a tag is found within all captions. This now means the program
can avoid reading all captions merely to see whether or not a tag should be deleted from the trie after
deleting one single instance of a tag within a caption.