feat: add autocomplete & tag suggestions
Trace user text entry to conduct live checks for existing tags in the dataset to suggest to them.
Add the SuggestBox class for autocompletion.
Add an autofill/suggestion box to the GUI, right of the tag entry field.

The autofill box references a trie data structure to track what tags exist in the current session, and as users type, suggests the first three words in the trie that it encounters with the prefix written by the user.
Users may also navigate between these three suggestions with the <Up> and <Down> arrow keys, if they wish.