feat: add dataset loading
Add tag management---deletion & addition of tags with dynamically-generated buttons. Add Queue and Trie data structures for the program. Add dataset loading features. Users may load a directory as a dataset with the existing 'Load' button. The program will search recursively for all '*.png' images within the directory and, for each that has no corresponding caption (a '*.TXT' file with an identical name), users will be prompted in queue on whether or not they would like to create one (selecting 'No' means that the *.png file will be ignored in the load).